requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
pillow==12.0.0
gunicorn==23.0.0
whitenoise==6.11.0
nepali-datetime==1.0.8.4
requests==2.32.5
openpyxl==3.1.5
xhtml2pdf==0.2.17
